/main 삭제, 지역 선택 너비 조정
